[refactor]: Nest Routes in Remittance page with Outlet
