+++ frontend/src/pages/Remittance/Payee.js
@@ -1,18 +1,16 @@
 import React from 'react';
-import TabBar from "../components/TabBar";
-import CustomHeading from "../components/CustomHeading";
-import {Box} from "@chakra-ui/react";
-import Page from "../components/Page";
+import TabBar from "../../components/TabBar";
+import CustomHeading from "../../components/CustomHeading";
 
 function Payee() {
     const tabs = ['Recent payees', 'My payees', 'New payees']
-    const panels = ['a','b','c']
+    const panels = ['a', 'b', 'c']
 
     return (
-        <Page>
+        <>
             <CustomHeading>Where would you like to send the money?</CustomHeading>
             <TabBar tabNames={tabs} tabPanels={panels}></TabBar>
-        </Page>
+        </>
     );
 }
